requirements.txt
uuid==1.30
python-dotenv==1.0.0
firebase-admin==6.2.0
flask==3.0.0